+++ module-system/module-implementations/sov-sequencer-registry/src/capabilities.rs
@@ -52,6 +52,19 @@ impl<GU: Gas> GasMeter<GU> for SequencerStakeMeter<GU> {
         Ok(())
     }
 
+    fn refund_gas(&mut self, gas: &GU) -> anyhow::Result<()> {
+        self.penalty_accumulator = self.penalty_accumulator.checked_sub(gas).ok_or_else(|| {
+            let gas_used = &self.penalty_accumulator;
+            anyhow::anyhow!(
+            "The gas to refund is greater than the gas used. The gas used is {gas_used}, the gas to refund is {gas}"
+        )})?;
+        self.remaining_stake = self
+            .remaining_stake
+            .saturating_add(gas.value(&self.gas_price));
+
+        Ok(())
+    }
+
     fn gas_used(&self) -> &GU {
         &self.penalty_accumulator
     }
@@ -148,36 +161,115 @@ impl<S: Spec, Da: sov_modules_api::DaSpec> SequencerRegistry<S, Da> {
 
 #[cfg(test)]
 mod tests {
-    use sov_modules_api::{GasArray, GasMeter, GasPrice, GasUnit};
+    use sov_modules_api::{Gas, GasArray, GasMeter, GasPrice, GasUnit};
 
-    use crate::SequencerStakeMeter;
+    use crate::{Amount, SequencerStakeMeter};
 
-    #[test]
-    fn test_sequencer_stake_meter_enough_gas() {
-        const INIT_STAKE: u64 = 100;
-        let mut stake_meter = SequencerStakeMeter {
-            remaining_stake: INIT_STAKE,
-            penalty_accumulator: GasUnit::<2>::ZEROED,
-            gas_price: GasPrice::<2>::from_slice(&[1; 2]),
-        };
-
-        stake_meter
-            .charge_gas(&GasUnit::<2>::from_slice(&[INIT_STAKE / 2; 2]))
-            .unwrap();
-        assert_eq!(stake_meter.remaining_stake, 0);
+    impl<GU: Gas> SequencerStakeMeter<GU> {
+        fn new(remaining_stake: Amount, gas_price: GU::Price) -> Self {
+            Self {
+                remaining_stake,
+                penalty_accumulator: GU::ZEROED,
+                gas_price,
+            }
+        }
     }
 
     #[test]
-    fn test_sequencer_stake_meter_not_enough_gas() {
-        const INIT_STAKE: u64 = 100;
-        let mut stake_meter = SequencerStakeMeter {
-            remaining_stake: INIT_STAKE,
-            penalty_accumulator: GasUnit::<2>::ZEROED,
-            gas_price: GasPrice::<2>::from_slice(&[1; 2]),
-        };
+    fn charge_gas_should_fail_if_not_enough_funds() {
+        let gas_price = GasPrice::<2>::from_slice(&[1; 2]);
 
-        stake_meter
-            .charge_gas(&GasUnit::<2>::from_slice(&[INIT_STAKE; 2]))
-            .unwrap_err();
+        let mut gas_meter = SequencerStakeMeter::new(0, gas_price.clone());
+
+        assert!(
+            gas_meter
+                .charge_gas(&GasUnit::<2>::from_slice(&[100; 2]))
+                .is_err(),
+            "The gas meter should not be able to charge gas if there is not enough funds"
+        );
+    }
+
+    #[test]
+    fn refund_gas_should_fail_if_not_enough_funds_consumed() {
+        let gas_price = GasPrice::<2>::from_slice(&[1; 2]);
+
+        let mut gas_meter = SequencerStakeMeter::new(100, gas_price.clone());
+
+        assert!(
+            gas_meter
+                .refund_gas(&GasUnit::<2>::from_slice(&[100; 2]))
+                .is_err(),
+            "The gas meter should not be able to refund gas if there is not enough gas consumed"
+        );
+    }
+
+    #[test]
+    fn try_charge_gas() {
+        const REMAINING_FUNDS: u64 = 100;
+        let gas_price = GasPrice::<2>::from_slice(&[1; 2]);
+
+        let mut gas_meter = SequencerStakeMeter::new(REMAINING_FUNDS, gas_price.clone());
+        assert!(
+            gas_meter
+                .charge_gas(&GasUnit::<2>::from_slice(&[REMAINING_FUNDS / 2; 2]))
+                .is_ok(),
+            "It should be possible to charge gas"
+        );
+        assert_eq!(
+            gas_meter.gas_used(),
+            &GasUnit::from_slice(&[REMAINING_FUNDS / 2; 2]),
+            "The gas used should be the same as the gas charged"
+        );
+        assert_eq!(gas_meter.gas_price(), &gas_price);
+        assert_eq!(
+            gas_meter.remaining_funds(),
+            0,
+            "There should be no more gas left in the meter"
+        );
+
+        assert!(
+            gas_meter
+                .charge_gas(&GasUnit::<2>::from_slice(&[1; 2]))
+                .is_err(),
+            "There should be no more gas left in the meter, hence charging more gas should fail"
+        );
+    }
+
+    #[test]
+    fn try_refund_gas() {
+        const REMAINING_FUNDS: u64 = 100;
+        let gas_price = GasPrice::from_slice(&[1; 2]);
+
+        let mut gas_meter = SequencerStakeMeter::new(REMAINING_FUNDS, gas_price);
+        assert!(
+            gas_meter
+                .charge_gas(&GasUnit::<2>::from_slice(&[REMAINING_FUNDS / 2; 2]))
+                .is_ok(),
+            "There should be enough gas left in the meter to charge"
+        );
+        assert_eq!(
+            gas_meter.remaining_funds(),
+            0,
+            "There should be no more gas left in the meter"
+        );
+
+        assert!(
+            gas_meter
+                .refund_gas(&GasUnit::from_slice(&[REMAINING_FUNDS / 4; 2]))
+                .is_ok(),
+            "Enough gas should have been consumed to be refunded",
+        );
+
+        assert_eq!(
+            gas_meter.gas_used(),
+            &GasUnit::from_slice(&[REMAINING_FUNDS / 4; 2],),
+            "The gas used amount should have decreased"
+        );
+
+        assert_eq!(
+            gas_meter.remaining_funds(),
+            REMAINING_FUNDS / 2,
+            "Half of the gas should be refunded"
+        );
     }
 }